新規陽性者における接触歴等不明者数: Rename components/cards/UntrackedRateCard.vue to components/index/CardsMonitoring/UntrackedRate/Card.vue
