+++ libraries/repositories/DailyPositiveDetailRepository.ts
@@ -21,7 +21,7 @@ export class DailyPositiveDetailRepository
    *
    * モニタリング項目(1) 新規陽性者数 (components/index/CardsMonitoring/MonitoringConfirmedCasesNumber/Card.vue)
    *
-   * モニタリング項目(3) 新規陽性者における接触歴等不明者数／増加比 (components/cards/UntrackedRateCard.vue)
+   * モニタリング項目(3) 新規陽性者における接触歴等不明者数／増加比 (components/index/CardsMonitoring/UntrackedRate/Card.vue)
    */
   constructor() {
     super(Convert.toDailyPositiveDetail(JSON.stringify(rawData)))
